chore: remove useTextColor composable
