+++ packages/ui/src/components/va-sidebar/VaSidebar.vue
@@ -23,7 +23,7 @@ import { computed, ref, shallowRef, watchEffect } from 'vue'
 
 import { VaConfig } from '../va-config'
 import { getGradientBackground } from '../../services/color'
-import { useColors, useTextColor, useBem, useClickOutside, useElementWidth, useComponentPresetProp } from '../../composables'
+import { useColors, useElementTextColor, useBem, useClickOutside, useElementWidth, useComponentPresetProp } from '../../composables'
 import { useSidebar } from './hooks/useSidebar'
 
 defineOptions({
@@ -100,7 +100,7 @@ watchEffect(() => {
 })
 
 const backgroundColorComputed = computed(() => getColor(props.color))
-const { textColorComputed } = useTextColor(backgroundColorComputed)
+const textColorComputed = useElementTextColor(backgroundColorComputed)
 
 const computedStyle = computed(() => {
   const backgroundColor = getColor(backgroundColorComputed.value)
