+++ packages/ui/src/composables/index.ts
@@ -33,7 +33,7 @@ export * from './useSize'
 export * from './useStateful'
 export * from './useSwipe'
 export * from './std/internal/useTemplateRef'
-export * from './useTextColor'
+export * from './useElementTextColor'
 export * from './useTimer'
 export * from './useTrackBy'
 export * from './useTranslation'
@@ -49,7 +49,6 @@ export * from './useGlobalConfig'
 export * from './usePlacementAliases'
 export * from './useDropdownable'
 export * from './std/internal/useTeleported'
-export * from './useElementTextColor'
 export * from './std/browser/useElementBackground'
 export * from './useNumericProp'
 export * from './useInputMask'
