--- packages/ui/src/services/vue-plugin/composables.ts
@@ -1,1 +1,0 @@
-export { useElementTextColor } from '../../composables/useElementTextColor'
